Reviews

+++ src/components/features/content/components/content-add-to-watch-list.tsx
@@ -4,19 +4,28 @@ import { BookmarkIcon } from 'lucide-react'
 
 import { useRouter } from 'next/navigation'
 
+import { SignUp } from '@/components/features/auth/sign-up'
 import { watchlistsAPI } from '@/redux/services/watchlists/watchlists.api'
+import { useDialog } from '@/utils/providers/dialog-provider'
 
 type ContentAddToWatchListProps = {
     contentId: number
+    isAuth: boolean
     isInWatchlist: boolean
 }
 
-export const ContentAddToWatchList = ({ contentId, isInWatchlist }: ContentAddToWatchListProps) => {
+export const ContentAddToWatchList = ({ contentId, isAuth, isInWatchlist }: ContentAddToWatchListProps) => {
     const router = useRouter()
+    const dialog = useDialog()
     const [addToWatchlist] = watchlistsAPI.useAddToWatchlistMutation()
     const [deleteFromWatchlist] = watchlistsAPI.useDeleteFromWatchlistMutation()
 
     const toggle = async () => {
+        if (!isAuth) {
+            dialog.open(<SignUp />)
+            return
+        }
+
         try {
             await (isInWatchlist ? deleteFromWatchlist(contentId) : addToWatchlist(contentId))
             router.refresh()
